+++ mccl/tools/utils.hpp
@@ -88,7 +88,7 @@ inline size_t d_gilbert_varshamov(size_t n, size_t k)
 
 inline int get_cryptographic_w(size_t n, size_t k)
 {
-  return ceil( 1.05 * double(d_gilbert_varshamov(n, k)) );
+  return int(ceil( 1.05 * double(d_gilbert_varshamov(n, k)) ));
 }
 
 MCCL_END_NAMESPACE
